server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
tensorflow==2.18.0
pillow==11.0.0
numpy==1.26.4
